Moving around functions to comply with requiremnets
